Breakdown side panel component
